refactor(minimax-h3): colocate relayout kernels
